test: cover lab patch capture audit findings
